feat: word frequency

+++ apps/backend/src/schemas/word.ts
@@ -1,4 +1,4 @@
-import { int, mysqlTable, varchar } from 'drizzle-orm/mysql-core';
+import { double, int, mysqlTable, varchar } from 'drizzle-orm/mysql-core';
 import { wordlistTable } from './wordlist';
 
 export const wordTable = mysqlTable('word', {
@@ -6,7 +6,7 @@ export const wordTable = mysqlTable('word', {
   word: varchar({ length: 255 }).notNull(),
   meaning: varchar({ length: 255 }),
   count: int().notNull(),
-  frequency: int().notNull(),
+  frequency: double().notNull(),
   wordlistId: int()
     .references(() => wordlistTable.id, { onDelete: 'cascade' })
     .notNull(),
